Replace legacy package.

+++ src/legacy/gui_summon.go
@@ -1,8 +1,5 @@
 package legacy
 
-/*
- */
-import "C"
 import (
 	"unsafe"
 
@@ -13,5 +10,5 @@ var (
 	Sub_4C26F0 func(a1 *gui.Window) int
 )
 
-//export sub_4C26F0
-func sub_4C26F0(a1 unsafe.Pointer) int { return Sub_4C26F0(AsWindowP(a1)) }
+// sub_4C26F0
+func sub_4C26F0(a1 unsafe.Pointer) int32 { return int32(Sub_4C26F0(AsWindowP(a1))) }
