+++ src/legacy/platform.go
@@ -1,9 +1,5 @@
 package legacy
 
-/*
-#include <stdbool.h>
-*/
-import "C"
 import "github.com/noxworld-dev/opennox-lib/platform"
 
 var (
@@ -11,14 +7,14 @@ var (
 	Nox_ticks_reset_416D40 func()
 )
 
-//export nox_platform_rand
-func nox_platform_rand() int { return platform.RandInt() }
+// nox_platform_rand
+func nox_platform_rand() int32 { return int32(platform.RandInt()) }
 
-//export nox_platform_srand
-func nox_platform_srand(seed C.uint) { platform.RandSeed(int64(seed)) }
+// nox_platform_srand
+func nox_platform_srand(seed uint32) { platform.RandSeed(int64(seed)) }
 
-//export nox_platform_get_ticks
-func nox_platform_get_ticks() C.uint { return C.uint(PlatformTicks()) }
+// nox_platform_get_ticks
+func nox_platform_get_ticks() uint32 { return uint32(PlatformTicks()) }
 
-//export nox_ticks_reset_416D40
+// nox_ticks_reset_416D40
 func nox_ticks_reset_416D40() { Nox_ticks_reset_416D40() }
